Move API endpoints into their own directory
