+++ backend/core/admin.py
@@ -1,9 +1,8 @@
 from django.contrib import admin
-from .models import User, Course, NonLectureSection, LectureSection, CourseEnrollment
+from .models import User, Course, NonLectureSection, LectureSection
 
 
 admin.site.register(User)
 admin.site.register(Course)
 admin.site.register(NonLectureSection)
 admin.site.register(LectureSection)
-admin.site.register(CourseEnrollment)
